Rename to pscompat.MessageMetadata
