+++ pubsublite/internal/wire/publish_batcher.go
@@ -18,7 +18,6 @@ import (
 	"errors"
 	"fmt"
 
-	"cloud.google.com/go/pubsublite/message"
 	"golang.org/x/xerrors"
 	"google.golang.org/api/support/bundler"
 	"google.golang.org/protobuf/proto"
@@ -29,7 +28,7 @@ import (
 var errPublishQueueEmpty = errors.New("pubsublite: received publish response from server with no batches in flight")
 
 // PublishResultFunc receives the result of a publish.
-type PublishResultFunc func(*message.Metadata, error)
+type PublishResultFunc func(*MessageMetadata, error)
 
 // messageHolder stores a message to be published, with associated metadata.
 type messageHolder struct {
@@ -142,7 +141,7 @@ func (b *publishMessageBatcher) OnPublishResponse(firstOffset int64) error {
 	batch, _ := frontElem.Value.(*publishBatch)
 	for i, msgHolder := range batch.msgHolders {
 		// Messages are ordered, so the offset of each message is firstOffset + i.
-		mm := &message.Metadata{Partition: b.partition, Offset: firstOffset + int64(i)}
+		mm := &MessageMetadata{Partition: b.partition, Offset: firstOffset + int64(i)}
 		msgHolder.onResult(mm, nil)
 		b.availableBufferBytes += msgHolder.size
 	}
