+++ pubsublite/internal/wire/resources.go
@@ -161,3 +161,17 @@ type subscriptionPartition struct {
 func (sp subscriptionPartition) String() string {
 	return fmt.Sprintf("%s/partitions/%d", sp.Path, sp.Partition)
 }
+
+// MessageMetadata holds properties of a message published to the Pub/Sub Lite
+// service.
+type MessageMetadata struct {
+	// The topic partition the message was published to.
+	Partition int
+
+	// The offset the message was assigned.
+	Offset int64
+}
+
+func (m *MessageMetadata) String() string {
+	return fmt.Sprintf("%d:%d", m.Partition, m.Offset)
+}
