+++ pubsublite/pscompat/message_test.go
@@ -142,3 +142,51 @@ func TestMessageTransforms(t *testing.T) {
 		})
 	}
 }
+
+func TestMessageMetadataStringEncoding(t *testing.T) {
+	for _, tc := range []struct {
+		desc    string
+		input   string
+		want    *MessageMetadata
+		wantErr bool
+	}{
+		{
+			desc:  "valid: zero",
+			input: "0:0",
+			want:  &MessageMetadata{Partition: 0, Offset: 0},
+		},
+		{
+			desc:  "valid: non-zero",
+			input: "3:1234",
+			want:  &MessageMetadata{Partition: 3, Offset: 1234},
+		},
+		{
+			desc:    "invalid: number",
+			input:   "1234",
+			wantErr: true,
+		},
+		{
+			desc:    "invalid: partition",
+			input:   "p:1234",
+			wantErr: true,
+		},
+		{
+			desc:    "invalid: offset",
+			input:   "10:9offset",
+			wantErr: true,
+		},
+	} {
+		t.Run(tc.desc, func(t *testing.T) {
+			got, gotErr := ParseMessageMetadata(tc.input)
+			if !testutil.Equal(got, tc.want) || (gotErr != nil) != tc.wantErr {
+				t.Errorf("ParseMessageMetadata(%q): got (%v, %v), want (%v, err=%v)", tc.input, got, gotErr, tc.want, tc.wantErr)
+			}
+
+			if tc.want != nil {
+				if got := tc.want.String(); got != tc.input {
+					t.Errorf("MessageMetadata(%v).String(): got %q, want: %q", tc.want, got, tc.input)
+				}
+			}
+		})
+	}
+}
